#27: Improved reporting when dapper methods cannot be found
