+++ Dapper.MoqTests/MockDatabase.Dapper.QueryAsync.cs
@@ -156,5 +156,15 @@ namespace Dapper.MoqTests
         /// </returns>
         public abstract Task<IEnumerable<T>> QueryAsync<T>(
             [ParameterType(ParameterType.CommandDefinition)] CommandDefinition command);
+
+        /// <summary>
+        /// Execute a query asynchronously using Task.
+        /// </summary>
+        /// <param name="cnn">The connection to query on.</param>
+        /// <param name="type">The type to return.</param>
+        /// <param name="command">The command used to query on this connection.</param>
+        public abstract Task<IEnumerable<T>> QueryAsync<T>(
+            [ParameterType(ParameterType.Type)] Type type,
+            [ParameterType(ParameterType.CommandDefinition)] CommandDefinition command);
     }
 }
