+++ Dapper.MoqTests/MockDatabase.Dapper.QuerySingle.cs
@@ -78,5 +78,15 @@ namespace Dapper.MoqTests
         /// <remarks>Note: the row can be accessed via "dynamic", or by casting to an IDictionary&lt;string,object&gt;</remarks>
         public abstract Task<T> QuerySingleAsync<T>(
             [ParameterType(ParameterType.CommandDefinition)] CommandDefinition command);
+
+        /// <summary>
+        /// Execute a single-row query asynchronously using Task.
+        /// </summary>
+        /// <param name="cnn">The connection to query on.</param>
+        /// <param name="command">The command used to query on this connection.</param>
+        /// <remarks>Note: the row can be accessed via "dynamic", or by casting to an IDictionary&lt;string,object&gt;</remarks>
+        public abstract Task<T> QuerySingleAsync<T>(
+            [ParameterType(ParameterType.Type)] Type type,
+            [ParameterType(ParameterType.CommandDefinition)] CommandDefinition command);
     }
 }
